refactor : 큰 의미 없는 기물 움직임 추상 클래스 삭제
